locking down delete item to owner

+++ backend/src/resolvers/Mutation.js
@@ -34,10 +34,17 @@ const Mutations = {
         }, info)
     },
 
-    async deleteItem(parent, args, ctx, info){
-        const where = {id: args.id};
-        const item = await ctx.db.query.item({where},`{id title}`);
-        return ctx.db.mutation.deleteItem({where}, info);
+    async deleteItem(parent, args, ctx, info) {
+        const where = { id: args.id };
+        const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission =>
+          ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+        if (!ownsItem && hasPermissions) {
+          throw new Error("You don't have permission to do that!");
+        }
+        return ctx.db.mutation.deleteItem({ where }, info);
     },
 
     async signup(parent, args, ctx, info){
